project/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
Pillow==10.1.0
python-dotenv==1.0.0
langchain-google-genai==1.0.8
langchain-core==0.1.52
langchain-groq==0.1.9
langchain-openai==0.1.20
PyMuPDF==1.23.26
sqlalchemy==2.0.23
pydantic==2.5.0
sqlite3